+++ source/extensions/octi.lab_tasks/octi/lab_tasks/tasks/locomotion/fetching/mdp/curriculums.py
@@ -15,7 +15,6 @@ import torch
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
-from omni.isaac.lab.assets import Articulation
 from omni.isaac.lab.managers import SceneEntityCfg
 from omni.isaac.lab.terrains import TerrainImporter
 
@@ -23,7 +22,7 @@ if TYPE_CHECKING:
     from omni.isaac.lab.envs import ManagerBasedRLEnv
 
 
-def terrain_levels_vel(
+def terrain_levels_pos(
     env: ManagerBasedRLEnv, env_ids: Sequence[int], asset_cfg: SceneEntityCfg = SceneEntityCfg("robot")
 ) -> torch.Tensor:
     """Curriculum based on the distance the robot walked when commanded to move at a desired velocity.
@@ -39,18 +38,19 @@ def terrain_levels_vel(
         The mean terrain level for the given environment ids.
     """
     # extract the used quantities (to enable type-hinting)
-    asset: Articulation = env.scene[asset_cfg.name]
     terrain: TerrainImporter = env.scene.terrain
-    command = env.command_manager.get_command("base_velocity")
+    # command = env.command_manager.get_command("base_velocity")
     # compute the distance the robot walked
-    distance = torch.norm(asset.data.root_pos_w[env_ids, :2] - env.scene.env_origins[env_ids, :2], dim=1)
+    if "goal_reached_count" not in env.extensions:
+        env.extensions["goal_reached_count"] = torch.zeros((env.num_envs,), dtype=torch.long, device=env.device)
+    score = env.extensions["goal_reached_count"][env_ids]
     # robots that walked far enough progress to harder terrains
-    move_up = distance > terrain.cfg.terrain_generator.size[0] / 2
+    move_up = score > 2
     # robots that walked less than half of their required distance go to simpler terrains
-    move_down = distance < torch.norm(command[env_ids, :2], dim=1) * env.max_episode_length_s * 0.5
+    move_down = score == 0
     move_down *= ~move_up
     # update terrain levels
     terrain.update_env_origins(env_ids, move_up, move_down)
+    env.extensions["goal_reached_count"][env_ids] = 0
     # return the mean terrain level
     return torch.mean(terrain.terrain_levels.float())
-