+++ source/extensions/octi.lab_tasks/octi/lab_tasks/tasks/locomotion/velocity/config/anymal_c/__init__.py
@@ -35,18 +35,6 @@ gym.register(
     },
 )
 
-gym.register(
-    id="Octi-Position-Rough-Anymal-C-v0",
-    entry_point="octi.lab.envs:OctiManagerBasedRLEnv",
-    disable_env_checker=True,
-    kwargs={
-        "env_cfg_entry_point": rough_env_cfg.AnymalCRoughPositionEnvCfg,
-        "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_rough_ppo_cfg.yaml",
-        "rsl_rl_cfg_entry_point": agents.rsl_rl_cfg.AnymalCRoughPositionPPORunnerCfg,
-        "skrl_cfg_entry_point": f"{agents.__name__}:skrl_rough_ppo_cfg.yaml",
-    },
-)
-
 
 
 gym.register(